refactor: infra 외부 어댑터를 external 패키지로 분리 (#439)
- auth/notification/sms/storage 어댑터를 infra.external.<concern>.<provider> 구조로 이동
- ExternalClientConfig import 경로 및 SharedBoundaryContractTest 경로 수정
- infra/README.md 전면 재작성 (bootstrap, 공개 표면, 가드 레일 등 11개 절)
