--- infra/src/main/java/com/beat/infra/auth/social/kakao/client/KakaoAuthApiClient.java
@@ -1,19 +1,0 @@
-package com.beat.infra.auth.social.kakao.client;
-
-import com.beat.infra.auth.social.kakao.response.KakaoAccessTokenResponse;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@FeignClient(name = "kakaoAuthApiClient", url = "https://kauth.kakao.com")
-public interface KakaoAuthApiClient {
-
-	@PostMapping(value = "/oauth/token", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-	KakaoAccessTokenResponse getOAuth2AccessToken(
-		@RequestParam("grant_type") String grantType,
-		@RequestParam("client_id") String clientId,
-		@RequestParam("redirect_uri") String redirectUri,
-		@RequestParam("code") String code
-	);
-}
