+++ infra/src/main/java/com/beat/infra/external/auth/social/kakao/response/KakaoUserResponse.java
@@ -1,0 +1,13 @@
+package com.beat.infra.external.auth.social.kakao.response;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record KakaoUserResponse(
+	@JsonProperty("id")
+	Long id,
+	@JsonProperty("connected_at")
+	String connectedAt,
+	@JsonProperty("kakao_account")
+	KakaoAccount kakaoAccount
+) {
+}
