+++ infra/src/main/java/com/beat/infra/external/notification/slack/vo/block/Block.java
@@ -1,0 +1,6 @@
+package com.beat.infra.external.notification.slack.vo.block;
+
+public sealed interface Block permits HeaderBlock, SectionBlock, DividerBlock {
+
+	String type();
+}
