+++ infra/src/main/java/com/beat/infra/external/notification/slack/vo/message/SlackMessage.java
@@ -1,0 +1,20 @@
+package com.beat.infra.external.notification.slack.vo.message;
+
+import com.beat.infra.external.notification.slack.vo.block.Block;
+import java.util.List;
+
+public record SlackMessage(
+	List<Attachment> attachments
+) {
+
+	public static SlackMessage newInstance(List<Block> blocks, String color) {
+		return new SlackMessage(List.of(Attachment.of(color, blocks)));
+	}
+
+	public record Attachment(String color, List<Block> blocks) {
+
+		public static Attachment of(String color, List<Block> blocks) {
+			return new Attachment(color, blocks);
+		}
+	}
+}
