+++ infra/src/main/java/com/beat/infra/external/notification/slack/vo/text/PlainText.java
@@ -1,0 +1,14 @@
+package com.beat.infra.external.notification.slack.vo.text;
+
+import static com.beat.infra.external.notification.slack.vo.SlackConstant.TEXT_TYPE_PLAIN;
+
+public record PlainText(
+	String type,
+	String text,
+	boolean emoji
+) implements Text {
+
+	public static PlainText newInstance(String text) {
+		return new PlainText(TEXT_TYPE_PLAIN, text, true);
+	}
+}
