+++ infra/src/main/java/com/beat/infra/external/notification/slack/vo/text/Text.java
@@ -1,0 +1,8 @@
+package com.beat.infra.external.notification.slack.vo.text;
+
+public sealed interface Text permits PlainText, MarkdownText {
+
+	String type();
+
+	String text();
+}
