+++ infra/src/main/java/com/beat/infra/external/storage/s3/S3InfraConfig.java
@@ -1,0 +1,37 @@
+package com.beat.infra.external.storage.s3;
+
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+@Configuration(proxyBeanMethods = false)
+public class S3InfraConfig {
+
+	@Value("${cloud.aws.credentials.access-key}")
+	private String accessKey;
+
+	@Value("${cloud.aws.credentials.secret-key}")
+	private String secretKey;
+
+	@Value("${cloud.aws.region}")
+	private String region;
+
+	@Bean
+	@Primary
+	public BasicAWSCredentials awsCredentialsProvider() {
+		return new BasicAWSCredentials(accessKey, secretKey);
+	}
+
+	@Bean
+	public AmazonS3 amazonS3(BasicAWSCredentials awsCredentials) {
+		return AmazonS3ClientBuilder.standard()
+			.withRegion(region)
+			.withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
+			.build();
+	}
+}
