--- infra/src/main/java/com/beat/infra/notification/slack/vo/SlackConstant.java
@@ -1,14 +1,0 @@
-package com.beat.infra.notification.slack.vo;
-
-public final class SlackConstant {
-
-	public static final String BRAND_COLOR = "#FF006B";
-	public static final String BLOCK_TYPE_HEADER = "header";
-	public static final String BLOCK_TYPE_SECTION = "section";
-	public static final String BLOCK_TYPE_DIVIDER = "divider";
-	public static final String TEXT_TYPE_PLAIN = "plain_text";
-	public static final String TEXT_TYPE_MARKDOWN = "mrkdwn";
-
-	private SlackConstant() {
-	}
-}
