--- infra/src/main/java/com/beat/infra/notification/slack/vo/block/Block.java
@@ -1,6 +1,0 @@
-package com.beat.infra.notification.slack.vo.block;
-
-public sealed interface Block permits HeaderBlock, SectionBlock, DividerBlock {
-
-	String type();
-}
