--- infra/src/main/java/com/beat/infra/notification/slack/vo/block/HeaderBlock.java
@@ -1,16 +1,0 @@
-package com.beat.infra.notification.slack.vo.block;
-
-import static com.beat.infra.notification.slack.vo.SlackConstant.BLOCK_TYPE_HEADER;
-
-import com.beat.infra.notification.slack.vo.text.PlainText;
-import com.beat.infra.notification.slack.vo.text.Text;
-
-public record HeaderBlock(
-	String type,
-	Text text
-) implements Block {
-
-	public static HeaderBlock newInstance(String text) {
-		return new HeaderBlock(BLOCK_TYPE_HEADER, PlainText.newInstance(text));
-	}
-}
