--- infra/src/main/java/com/beat/infra/notification/slack/vo/text/Text.java
@@ -1,8 +1,0 @@
-package com.beat.infra.notification.slack.vo.text;
-
-public sealed interface Text permits PlainText, MarkdownText {
-
-	String type();
-
-	String text();
-}
